make NFluxNeutrinos() return a long int
git-svn-id: file:///home/mroda/Software/Genie/NewGlobal/trunk@1899 cc9776de-3512-45ca-aafc-e2d9ed43c22c

+++ src/EVGDrivers/GMCJDriver.h
@@ -59,8 +59,8 @@ public :
   EventRecord * GenerateEvent (void);
 
   // methods for enquiring info needed for computing the generated sample normalization
-  double GlobProbScale  (void) const { return fGlobPmax;       }
-  double NFluxNeutrinos (void) const { return fNFluxNeutrinos; }
+  double   GlobProbScale  (void) const { return fGlobPmax;                  }
+  long int NFluxNeutrinos (void) const { return (long int) fNFluxNeutrinos; }
 
 private:
